move deprecate_ndim_indexing
